Update test namespaces in System.Threading.Tasks.Parallel, per #2898 .  - Clean up `using`s.  - Some types moved inside existing namespace.  - ParallelStateTest only updated to reference moved logger (indentation changes too large).
