+++ src/System.Threading.Tasks.Parallel/tests/Logger.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-namespace CoreFXTestLibrary
+namespace System.Threading.Tasks.Tests
 {
     // A dummy class for passing through System.Console calls until the System.Console contract is available.
     // To-do: Remove this workaround when possible.
